moved session and session_manager to models folder
