+++ app/routes/courses_route.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request, make_response, Response
 from flask.json import jsonify
 from app.models.courses import Courses
 from app.models.departments import Departments
-from app.SessionManager import SessionManager
+from app.models.session_manager import SessionManager
 
 SManager = SessionManager()
 app_course_routes = Blueprint('courses_routes', __name__)
