+++ app/routes/degrees_routes.py
@@ -2,7 +2,7 @@ from flask import Blueprint, request, make_response, Response
 from flask.json import jsonify
 from app.models.degrees import Degrees
 from app.models.departments import Departments
-from app.SessionManager import SessionManager
+from app.models.session_manager import SessionManager
 
 SManager = SessionManager()
 app_degrees_routes = Blueprint('degrees_routes', __name__)
