+++ app/routes/history_route.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, make_response, Response
 from flask.json import jsonify
 from app.models.history import History
-from app.SessionManager import SessionManager
+from app.models.session_manager import SessionManager
 
 SManager = SessionManager()
 app_history_routes = Blueprint('history_routes', __name__)
